requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly.express
scikit-learn
streamlit_extras
streamlit_option_menu